wallet: extract CreateDerivedAccount shared workflow
